// src/openapi/model/api-mypage-credit-card.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mtg_server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 *
 * @export
 * @interface ApiMypageCreditCard
 */
export interface ApiMypageCreditCard {
    /**
     *
     * @type {string}
     * @memberof ApiMypageCreditCard
     */
    cardNumber?: string;
    /**
     *
     * @type {string}
     * @memberof ApiMypageCreditCard
     */
    expiration?: string;
    /**
     *
     * @type {string}
     * @memberof ApiMypageCreditCard
     */
    securityCode?: string;
    /**
     *
     * @type {string}
     * @memberof ApiMypageCreditCard
     */
    token?: string;
    /**
     *
     * @type {string}
     * @memberof ApiMypageCreditCard
     */
    tokenJsUrl?: string;
    /**
     *
     * @type {string}
     * @memberof ApiMypageCreditCard
     */
    shopId?: string;
}